requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.23.0
attrs==17.3.0
autobahn==17.10.1
Automat==0.6.0
blessings==1.6
bpython==0.16
certifi==2017.11.5
cffi==1.11.2
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.4
curtsies==0.2.11
daphne==1.3.0
Django==1.9
django-allauth==0.34.0
django-filter==1.1.0
django-rest-auth==0.9.2
djangorestframework-jwt==1.11.0
enum34==1.1.6
greenlet==0.4.12
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipaddress==1.0.18
Markdown==2.6.10
msgpack-python==0.4.8
ndg-httpsclient==0.4.3
oauthlib==2.0.6
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.5.3
pyOpenSSL==17.5.0
python-openid==2.2.5
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.44.0
zope.interface==4.4.3
